Simplified integration test
