+++ plugin/src/test/test-fixtures/multi-module/app/src/androidTest/java/org/neotech/app/multimoduleapplication/TouchLibraryCodeTest.kt
@@ -1,0 +1,23 @@
+package org.neotech.app.multimoduleapplication
+
+import android.support.test.runner.AndroidJUnit4
+
+import org.junit.Test
+import org.junit.runner.RunWith
+
+import org.neotech.library.android.LibraryAndroidJava
+import org.neotech.library.android.LibraryAndroidKotlin
+
+@RunWith(AndroidJUnit4::class)
+class TouchLibraryCodeTest {
+
+    @Test
+    fun touchAndroidLibraryCodeJava() {
+        LibraryAndroidJava.getInstance().getName2()
+    }
+
+    @Test
+    fun touchAndroidLibraryCodeKotlin() {
+        LibraryAndroidKotlin.getName2()
+    }
+}
