+++ plugin/src/test/test-fixtures/multi-module/app/src/main/java/org/neotech/app/MainActivity.kt
@@ -1,0 +1,17 @@
+package org.neotech.app
+
+import android.annotation.SuppressLint
+import android.support.v7.app.AppCompatActivity
+import android.os.Bundle
+import android.widget.TextView
+import org.neotech.library.android.LibraryAndroidKotlin
+
+class MainActivity : AppCompatActivity() {
+
+    @SuppressLint("SetTextI18n")
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+        findViewById<TextView>(R.id.text).text = "${LibraryAndroidKotlin.getName()}"
+    }
+}
