--- plugin/src/test/test-fixtures/multi-module/library_a/src/main/java/org/neotech/library/a/LibraryAJava.java
@@ -1,21 +1,0 @@
-package org.neotech.library.a;
-
-public final class LibraryAJava {
-
-    private static LibraryAJava INSTANCE;
-
-    public static synchronized LibraryAJava getInstance() {
-        if(INSTANCE == null) {
-            INSTANCE = new LibraryAJava();
-        }
-        return INSTANCE;
-    }
-
-    private LibraryAJava() {
-
-    }
-
-    public String getName() {
-        return "LibraryAJava";
-    }
-}
