--- plugin/src/test/test-fixtures/multi-module/library_a/src/main/java/org/neotech/library/a/LibraryAKotlin.kt
@@ -1,9 +1,0 @@
-package org.neotech.library.a
-
-object LibraryAKotlin {
-
-    @JvmStatic
-    fun getName(): String {
-        return "LibraryAKotlin"
-    }
-}
