+++ plugin/src/test/test-fixtures/multi-module/library_android/build.gradle
