+++ plugin/src/test/test-fixtures/multi-module/library_android/proguard-rules.pro
