+++ plugin/src/test/test-fixtures/multi-module/library_android/src/main/AndroidManifest.xml
