--- plugin/src/test/test-fixtures/multi-module/library_b/.gitignore
@@ -1,1 +1,0 @@
-/build
