--- plugin/src/test/test-fixtures/multi-module/library_b/src/main/AndroidManifest.xml
@@ -1,2 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.neotech.library.b" />
