--- plugin/src/test/test-fixtures/multi-module/library_b/src/main/java/org/neotech/library/b/LibraryBJava.java
@@ -1,21 +1,0 @@
-package org.neotech.library.b;
-
-public final class LibraryBJava {
-
-    private static LibraryBJava INSTANCE;
-
-    public static synchronized LibraryBJava getInstance() {
-        if(INSTANCE == null) {
-            INSTANCE = new LibraryBJava();
-        }
-        return INSTANCE;
-    }
-
-    private LibraryBJava() {
-
-    }
-
-    public String getName() {
-        return "LibraryBJava";
-    }
-}
