--- plugin/src/test/test-fixtures/multi-module/library_b/src/main/java/org/neotech/library/b/LibraryBKotlin.kt
@@ -1,9 +1,0 @@
-package org.neotech.library.b
-
-object LibraryBKotlin {
-
-    @JvmStatic
-    fun getName(): String {
-        return "LibraryBKotlin"
-    }
-}
