--- plugin/src/test/test-fixtures/multi-module/library_b/src/main/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-    <string name="app_name">Library B</string>
-</resources>
