+++ plugin/src/test/test-fixtures/multi-module/library_java/.gitignore
