+++ plugin/src/test/test-fixtures/multi-module/library_java/src/main/java/org/neotech/library/java/Test.java
